Run getframe() and _ProcessGuiEvents() in the main loop thread.
Using code from Stephen Langer, we can cause the thread executing
the worksheet to send a callback to the main loop thread and wait
for its execution.  Thus, we can call getframe from the main loop
thread, but still pass the result back to screenshot or record in
the worksheet.  So these work now.

Well, sorta.  They need the _widget to have been realized so that
they can scrape the pixels.  Thus, these sorts of statements can only
be executed after the widget has been created.  It's a bit clunky,
but it may be the best we can do.