Ensure getOpenGlInfo is only called at safe times
visvis.misc.getOpenGlInfo() can only be called safely after an
OpenGL context has been realized.  Otherwise, it may segfault.
Some functions (like volshow) call this to determine the system's
OpenGL capabilities, but in revis this call may occur before the
OpenGL context is created.  This function only needs to be called
once, as its output is cached.  So we call it as soon as safely
possible and prevent it from being called in other places.

This provides a partial solution to issue #2, although it means
that a volshow call in the first figure created in a session will
use volshow2, even if you system supports volshow3.