Be able to grab focus and capture keypress events
Changes in visvis allow control of the view through keypress events.
To get those, we have to grab focus when the figure is clicked.  Also,
be sure to return True after the keypress event, so that it doesn't
bubble up to the TextView and cause a character to be inserted.

Also changed to viewParams display due to updates in visvis.